requirements.txt
discord.py>=2.4,<3.0
SQLAlchemy[asyncio]>=2.0,<3.0
aiosqlite>=0.20
alembic>=1.13
httpx>=0.27
APScheduler>=3.10,<4.0
python-dotenv>=1.0
